gpg: New option --quick-keytocard
* wrote the documentation for the new option (untested)
* g10/card-util.c (send_keytocard) added
* g10/gpg.c added aQuickKeyToCard to cmd_and_opt_values
* g10/gpg.c added ARGPARSE_c with aQuickKeyToCard
* g10/gpg.c added handling for aQuickKeyToCard
* g10/keyedit.c (get_keyno_from_slot_usage) new function
maps commandline option to keyno.
* g10/keyedit.c (is_keyno_matching_usage) tests if the
key has the right usage for the keyno.
* g10/keyedit.c (keyedit_quick_keytocard) new function
to prepare the commandline parameter to call
(send_keytocard) to transfer the key.
* g10/main.h (keyedit_quick_keytocard) forward decl
* g10/main.h (send_keytocard) forward decl

Tested: by hand i send ~15 keys to reseted yubikeys
and try successfully to use the key for sign,encr,auth
Missing: in batch we need to pass the adminpin via
loopback pinentry

Signed-off-by: Meno Abels <meno.abels@adviser.com>