gpg: Enable to pass multiple passphrases
This feature is required to copy keys in a smartcard without
any user interaction.
It currently only used by --quick-keytocard with a command line
like this:
gpg --batch --pinentry-mode loopback --passphrase-file ~/p1
--passphrase-file ~/p2 --quick-keytocard <keyid> auth

This is not a designed solution(hack). This could break any
time.
Next I will try to build tests explain the odd behavior.
The problem is (get_static_passphrase) now have a side
effect that after the call to (get_static_passphrase) the
next password will be returned on the next call to
(get_static_passphrase). And (get_static_passphrase) the
order and of calls defines the returned password.

I discovered that in gpgsm almost the same code is used.
But it is in different files. So should id refactor this?
Currently i did not changed this gpgsm.

* g10/gpg.c changed switch cases oPassphraseFD/oPassphraseFile
to store the incoming file descriptor.
* g10/gpg.c passed the collected fds to the new
(read_passphrase_from_fds)
* g10/keydb.h added STATIC_PASSWORD_COUNT constant (4)
* g10/keydb.h modified (read_passphrase_from_fd) to
(read_passphrase_from_fds) to pass multiple fds
* g10/passphrase.c removed fd_passwd
* g10/passphrase.c added fd_passwds[STATIC_PASSWORD_COUNT]
* g10/passphrase.c added fd_passwd_idx use count
* g10/passphrase.c added fd_passwd_req roundrobin counter
* g10/passphrase.c (have_static_passphrase) uses fd_passwd_idx
* g10/passphrase.c (get_static_passphrase) roundrobin behavior
* g10/passphrase.c (set_passphrase_from_string) appends to fd_passwds
* g10/passphrase.c (read_passphrase_from_fd) static and wipe added
* g10/passphrase.c (read_passphrase_from_fds) loops over fds
* g10/passphrase.c (passphrase_to_dek) change direct use of
fw_passwd to (get_static_passphrase)

Signed-off-by: Meno Abels <meno.abels@adviser.com>